requirements.txt
amqp==5.0.6
beautifulsoup4==4.9.3
blinker==1.4
celery==5.2.2
docutils==0.17.1
emoji==1.6.1
flask-cors==3.0.10
Flask-Dance[sqla]==6.0.0
flask-login==0.6.0
Flask-SQLAlchemy==2.5.1
Flask==2.1.1
gevent==21.1.2
gunicorn==20.1.0
language-detector>=5.0.2
mistune==2.0.0rc1
mysqlclient>=2.1.0
PyGithub==1.55
PyGorse>=0.4.1
python-dateutil==2.8.1
python-logging-loki~=0.3.1
redis==4.2.2
requests~=2.27.1
werkzeug~=2.0.3
